KULRICE-1780
I removed the circular dependancies by moving the KEW Platform implementation into the Rice Core module. This is now a forced change because the packages have changed from edu.iu.edu... to org.kuali... and these were used in configuration. All configuration in the project has been changed, but if you use this in outside configuration, you must change to something like:

	<param name="datasource.platform">org.kuali.rice.database.platform.OraclePlatform</param>

There is still work to do to merge this Plaform with KualiDBPlatform and the KEN Platform.