KULRICE-2949: Create a proper Rice "portal" https://test.kuali.org/jira/browse/KULRICE-2949
Created a new configuration property, portal.show.sample.app, which needs to be set to "true" to show the Sample Travel App links in the portal.  Due to the way the configuration properties are accessed in the JSPs, the type for the property value had to be a String, which is a bit strange.