KULRICE-5678 Third party dependency licensing is now embedded in all jar's at META-INF/THIRD-PARTY.txt.
Duplicate licenses are condensed under the correct license name.  Reduced licenses from 60+ down to 23.

Jenkins uses a plugin execution to fail a build job should a dependency be added to the build tree without a known license.

Process is in place to make this easy to keep up to date so it no longer requires a bunch of work at release time.