Changes to WebUtils.java#filterHtmlAndReplaceRiceMarkup(..) to allow passing of URLs (in support of KFS req for KNS question framework).
Added 3 new filter patterns. One to remove value when 'javascript' is passed inside the [a <value>] tag to avoid XSS issues. One to deal with <a href="<value"> tag, and one to deal with </a> tag.

To pass a URL and have it properly processed, it should look like this:
"This is an example question which would link to [a http://www.google.com]Google.com[/a]".

Also changed the pattern map from a HashMap to a LinkedHashMap since we'd want the javascript pattern to be matched before a href tag is created, otherwise it's useless. 

Fixes KULRICE-8417 (https://jira.kuali.org/browse/KULRICE-8417)