KULRICE-12855 IT Failure JpaPersistenceProviderTest testFindWithResultsWindow expected:<[mcs0epVGW4]> but was:<[FK5z4pBhjm]>
Issue:
After doing more testing, determined the test was attempting to order on a column of varchar and not a number/integer  which would not insure ascending order similar to the objects list.  Another issue was found with the QueryTranslator.  The translator was not preserving "order by" clauses when translating to a query.

Fix:
The IT was updated to pull the persisted objects back from the data store in ascending order by the number column, a second query is used to get the results window.  The results window is now compared against the first query rather than the objects list as it was previously doing.  A new test method was also created to test the ascending and descending order by.