KULRICE-12855 IT Failure JpaPersistenceProviderTest testFindWithResultsWindow expected:<[mcs0epVGW4]> but was:<[FK5z4pBhjm]>
After first code review, attempted to update OrderBy test to use mulitple order by clauses.  The test was failing.  The implementing method addOrderBy in the NativeJpaQueryTranslator class needed to pull the existing list of order bys and add the latest to the existing list.  currently it was overriding the current clause.