KULRICE-12946 IT Failure DataDictionarySearchableAttributeTest
Not able to reproduce the errors in the Jira.  I have fixed another error because of an update to the SqlBuilder class. It no longer accepts empty values "!".  I removed one test from the testWildcardsAndMultiSelectsOnDDSearchableAttributes so the new exception would not be thrown.  I also remove an if statement from the assertDDSearchableAttributeWildcardsWork method that did not do anything.