KULRICE-12974 AFT Failure ConfigComponentCreateNewBkMrkAft Deadlock found when trying to get lock
The namespace code and the component code are the unique identifiers for components.  The AFT was using the same namespace and component code across all tests in this class.  After the first test would create the component any other test trying to create the same component would fail on a duplicate error.  I updated the testConfigComponentCreateNewDuplicateEntry method within the ConfigComponentCreateNewAftBase class to concatenate the tempValue parameter onto the component id value to have unique component ids per method call.