KULRICE-13206
Removed the disabling of the paste button from the disableButtons JS file.  This was causing the button to be disabled before the RuleMaintenance paste action script code was executed.  I also added a disablePasteButton method call within the RuleMaintenance paste action script code so that after the paste button is clicked it is disabled.