KULRICE-14183: RouteLog re-ordering of action requests for primary delegates incorrectly persists to the database
Problem:
RouteLogAction#switchActionRequestPositionsIfPrimaryDelegatesPresent was
re-ordering the generated action requests if Primary delegates were
present and was persisting the change to the database. This behavior is
not desired as the action requests hierarchy was being modified
incorrectly.
Fix: Detached the action request object so that the re-ordering of
action requests in switchActionRequestPositionsIfPrimaryDelegatesPresent
method would not be saved.