RICE-7 - ported 76648f4238f9a44faf8772b05105925680ee19a2 from rice-2.5
Added the ability to completely disable LoadTime Weaving for JPA
This was due to the fact that when using Tomcat 8 (with it's new classloader which is natively instrumentable) without using the TomcatInstrumentableClassLoader from the spring-instrument-tomcat library, then even when working with compiled Rice jars that were statically weaved, a "java.lang.ClassCircularityError: org/apache/log4j/spi/ThrowableInformation" exception would be thrown on startup.

The solution is to either use the spring-instrument-tomcat library still (and set up the META-INF/context.xml in your WAR file to configure "<Loader loaderClass="org.springframework.instrument.classloading.tomcat.TomcatInstrumentableClassLoader"/>" or simply disable load time weaving entirely.

With this commit, loadtime weaving is disabled by default and has to be explicitly enabled when needed or desired using the following configuration parameter:

<param name="rice.jpa.ltw.enabled">true</param>